Implement map class
This is to represent the entire problem. We can generate new random
instances of the problem by calling Map.generate. This will random place
26 cities on a 100 x 100 grid and connect them using the R2 method
described in the problem description. That is, for each city it will
find the 5 closest cities and connect it to 3 of those. The distance
between the cities is the Euclidian distance between them based on their
coordinates.