Implement the Zero strategy
Although this is not a particularly interesting example, this strategy
simply returns 0 for the heuristic function for any node. The project
asked us to implement greedy and A* search using 2 different heuristics.
The obvious choice for the first was Euclidian distance to goal, but the
second is less obvious. Since our cities are on a grid, any heuristic
function that returns a value greater than the straight line distance is
not admissable. Since we discussed manhattan distance in class, I assume
that was the intended second heuristic. However, Manhattan distance is
clearly not admissable for this particular problem, so I decided to go
with the zero function.